+++ packages/medusa-plugin-add-ons/src/api/routes/store/get-by-product.js
@@ -13,7 +13,13 @@ export default async (req, res) => {
 
   try {
     const addOnService = req.scope.resolve("addOnService")
-    const addOn = await addOnService.retrieveByProduct(value.product_id)
+    let addOn = await addOnService.retrieveByProduct(value.product_id)
+    addOn = await addOnService.decorate(
+      addOn,
+      ["name", "valid_for", "prices"],
+      ["valid_for"]
+    )
+
     res.json({ add_on: addOn })
   } catch (err) {
     throw err