+++ packages/medusa-plugin-mailchimp/src/services/__tests__/mailchimp.js
@@ -1,0 +1,21 @@
+const MailchimpMock = jest.fn().mockImplementation(() => {
+  return {
+    subscribeNewsletter: jest.fn().mockReturnValue(Promise.resolve("success")),
+  }
+})
+
+describe("MailchimpService", () => {
+  describe("newsletterSubscribe", () => {
+    let result
+    beforeAll(async () => {
+      jest.clearAllMocks()
+      const mailchimpService = new MailchimpMock()
+
+      result = await mailchimpService.subscribeNewsletter("medusa@medusa.com")
+    })
+
+    it("resolves successfully", () => {
+      expect(result).toEqual("success")
+    })
+  })
+})