+++ packages/medusa/src/models/schemas/payment-method.js
@@ -4,6 +4,6 @@
 import mongoose from "mongoose"
 
 export default new mongoose.Schema({
-  provider_id: { type: String, required: true },
+  provider_id: { type: String },
   data: { type: mongoose.Schema.Types.Mixed, default: {} },
 })