+++ packages/medusa/src/services/__mocks__/discount.js
@@ -40,7 +40,10 @@ export const DiscountServiceMock = {
     })
   }),
   list: jest.fn().mockImplementation(data => {
-    return Promise.resolve([])
+    return Promise.resolve([{}])
+  }),
+  decorate: jest.fn().mockImplementation(data => {
+    return Promise.resolve(data)
   }),
   addRegion: jest.fn().mockReturnValue(Promise.resolve()),
   removeRegion: jest.fn().mockReturnValue(Promise.resolve()),