+++ packages/medusa/src/services/__mocks__/fulfillment-provider.js
@@ -13,11 +13,47 @@ export const DefaultProviderMock = {
 
     return Promise.resolve(false)
   }),
+  cancelFulfillment: jest.fn().mockImplementation(data => {
+    return {}
+  }),
   calculatePrice: jest.fn().mockImplementation(data => {
     return Promise.resolve()
   }),
   createOrder: jest.fn().mockImplementation(data => {
     return Promise.resolve(data)
+  }),
+  getFulfillmentDocuments: jest.fn().mockImplementation(() => {
+    return Promise.resolve([
+      {
+        name: "Test",
+        type: "pdf",
+        base_64: "verylong",
+      },
+    ])
+  }),
+  createReturn: jest.fn().mockImplementation(data => {
+    return Promise.resolve({
+      ...data,
+      shipped: true,
+    })
+  }),
+  getReturnDocuments: jest.fn().mockImplementation(() => {
+    return Promise.resolve([
+      {
+        name: "Test Return",
+        type: "pdf",
+        base_64: "verylong return",
+      },
+    ])
+  }),
+  getShipmentDocuments: jest.fn().mockImplementation(() => {
+    return Promise.resolve([
+      {
+        name: "Test Shipment",
+        type: "pdf",
+        base_64: "verylong shipment",
+      },
+    ])
   }),
 }