+++ packages/medusa/src/services/cart.js
@@ -246,7 +246,9 @@ class CartService extends BaseService {
     if (expandFields.includes("region")) {
       c.region = await this.regionService_.retrieve(cart.region_id)
     }
-    return c
+
+    const final = await this.runDecorators_(c)
+    return final
   }
 
   /**
@@ -375,7 +377,7 @@ class CartService extends BaseService {
 
     // If content matches one of the line items currently in the cart we can
     // simply update the quantity of the existing line item
-    if (currentItem) {
+    if (currentItem && validatedLineItem.should_merge) {
       const newQuantity = currentItem.quantity + validatedLineItem.quantity
 
       // Confirm inventory
@@ -1106,7 +1108,7 @@ class CartService extends BaseService {
    * @param {string} regionId - the id of the region to set the cart to
    * @return {Promise} the result of the update operation
    */
-  async setRegion(cartId, regionId) {
+  async setRegion(cartId, regionId, countryCode) {
     const cart = await this.retrieve(cartId)
     const region = await this.regionService_.retrieve(regionId)
 
@@ -1120,6 +1122,7 @@ class CartService extends BaseService {
       const newItems = await Promise.all(
         cart.items.map(async lineItem => {
           try {
+            lineItem.has_shipping = false
             lineItem.content = await this.updateContentPrice_(
               lineItem.content,
               region._id
@@ -1134,17 +1137,28 @@ class CartService extends BaseService {
       update.items = newItems.filter(i => !!i)
     }
 
-    // If the country code of a shipping address is set we need to clear it
     let shippingAddress = cart.shipping_address || {}
-    if (!_.isEmpty(shippingAddress) && shippingAddress.country_code) {
-      shippingAddress.country_code = ""
+    if (countryCode !== undefined) {
+      if (!region.countries.includes(countryCode)) {
+        throw new MedusaError(
+          MedusaError.Types.NOT_ALLOWED,
+          `Country not available in region`
+        )
+      }
+      shippingAddress.country_code = countryCode
       update.shipping_address = shippingAddress
-    }
+    } else {
+      // If the country code of a shipping address is set we need to clear it
+      if (!_.isEmpty(shippingAddress) && shippingAddress.country_code) {
+        shippingAddress.country_code = ""
+        update.shipping_address = shippingAddress
+      }
 
-    // If there is only one country in the region preset it
-    if (region.countries.length === 1) {
-      shippingAddress.country_code = region.countries[0]
-      update.shipping_address = shippingAddress
+      // If there is only one country in the region preset it
+      if (region.countries.length === 1) {
+        shippingAddress.country_code = region.countries[0]
+        update.shipping_address = shippingAddress
+      }
     }
 
     // Shipping methods are determined by region so the user needs to find a