+++ packages/medusa/src/services/user.js
@@ -231,7 +231,7 @@ class UserService extends BaseService {
   async setPassword(userId, password) {
     const user = await this.retrieve(userId)
 
-    const hashedPassword = await bcrypt.hash(password, 10)
+    const hashedPassword = await this.hashPassword_(password)
     if (!hashedPassword) {
       throw new MedusaError(
         MedusaError.Types.DB_ERROR,
@@ -278,7 +278,8 @@ class UserService extends BaseService {
   async decorate(user, fields, expandFields = []) {
     const requiredFields = ["_id", "metadata"]
     const decorated = _.pick(user, fields.concat(requiredFields))
-    return decorated
+    const final = await this.runDecorators_(decorated)
+    return final
   }
 
   /**