+++ www/gatsby-browser.js
@@ -1,0 +1,6 @@
+import React from "react"
+import { ThemeProvider as Provider } from "./src/theme"
+
+export const wrapPageElement = ({ element }) => {
+  return <Provider>{element}</Provider>
+}