+++ www/src/theme/spacing.js
@@ -1,0 +1,1 @@
+export default [0, 4, 8, 16, 32, 64]