+++ packages/medusa/CHANGELOG.md
@@ -3,6 +3,160 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.35](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.34...@medusajs/medusa@1.0.35) (2020-11-02)
+
+
+### Bug Fixes
+
+* add country code on preCartCreation ([e3faa64](https://github.com/medusajs/medusa/commit/e3faa646a7b91d2686b56718df46bea5b709731b))
+
+
+
+
+
+## [1.0.34](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.33...@medusajs/medusa@1.0.34) (2020-10-31)
+
+
+### Bug Fixes
+
+* adds field selection to list order endpoint ([#133](https://github.com/medusajs/medusa/issues/133)) ([b1786ce](https://github.com/medusajs/medusa/commit/b1786ce9d8684b1611c994380e39b25c3ac181a1))
+
+
+
+
+
+## [1.0.33](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.32...@medusajs/medusa@1.0.33) (2020-10-28)
+
+
+### Features
+
+* **medusa:** adds support for gift cards ([#132](https://github.com/medusajs/medusa/issues/132)) ([f2c62cd](https://github.com/medusajs/medusa/commit/f2c62cd2321c9013c15160a80598f912daef4647))
+
+
+
+
+
+## [1.0.32](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.31...@medusajs/medusa@1.0.32) (2020-10-20)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.31](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.30...@medusajs/medusa@1.0.31) (2020-10-20)
+
+
+### Features
+
+* **medusa-interfaces:** Adds schema options to base model ([cc23a3b](https://github.com/medusajs/medusa/commit/cc23a3b0706c41ec57bb25ea3de9c6e39bd04f31))
+
+
+
+
+
+## [1.0.30](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.29...@medusajs/medusa@1.0.30) (2020-10-20)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.29](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.28...@medusajs/medusa@1.0.29) (2020-10-20)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.28](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.27...@medusajs/medusa@1.0.28) (2020-10-19)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.27](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.26...@medusajs/medusa@1.0.27) (2020-10-17)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.26](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.25...@medusajs/medusa@1.0.26) (2020-10-14)
+
+
+### Features
+
+* return shipping and flow ([#125](https://github.com/medusajs/medusa/issues/125)) ([c1e821d](https://github.com/medusajs/medusa/commit/c1e821d9d4d33756c7309e5cf110d7aa9b67297d))
+
+
+
+
+
+## [1.0.25](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.24...@medusajs/medusa@1.0.25) (2020-10-08)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.24](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.23...@medusajs/medusa@1.0.24) (2020-10-08)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.23](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.22...@medusajs/medusa@1.0.23) (2020-10-06)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.22](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.21...@medusajs/medusa@1.0.22) (2020-10-06)
+
+**Note:** Version bump only for package @medusajs/medusa
+
+
+
+
+
+## [1.0.21](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.20...@medusajs/medusa@1.0.21) (2020-10-05)
+
+
+### Bug Fixes
+
+* **medusa-plugin-brightpearl:** reliable oauth ([#119](https://github.com/medusajs/medusa/issues/119)) ([0889059](https://github.com/medusajs/medusa/commit/0889059ba178e976c5f6c4a0e938a463dde29554))
+
+
+
+
+
+## [1.0.20](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.19...@medusajs/medusa@1.0.20) (2020-10-05)
+
+
+### Bug Fixes
+
+* **medusa-plugin-addon:** Fixes admin endpoints, Adds flag to avoid merging add-on line-items ([d8483cd](https://github.com/medusajs/medusa/commit/d8483cd1352ecc587112723786b7c31882f9416e))
+
+
+### Features
+
+* webshipper ([#118](https://github.com/medusajs/medusa/issues/118)) ([893a7f6](https://github.com/medusajs/medusa/commit/893a7f69afea67e854a67fc3b92c8a10c9c1b75c))
+
+
+
+
+
 ## [1.0.19](https://github.com/medusajs/medusa/compare/@medusajs/medusa@1.0.18...@medusajs/medusa@1.0.19) (2020-09-21)
 
 **Note:** Version bump only for package @medusajs/medusa